Move constants to app/constants
