+++ app/components/Base/StatusText.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Text from './Text';
 import { colors } from '../../styles/common';
 import { StyleSheet } from 'react-native';
-import { FIAT_ORDER_STATES } from '../UI/FiatOrders/utils';
+import { FIAT_ORDER_STATES } from '../../constants/on-ramp';
 import { strings } from '../../../locales/i18n';
 
 const styles = StyleSheet.create({
